requirements.txt
Flask
Pillow
watchdog
